Move StoredRegex.cs and split PatternItem.cs into Text-Grab.Core
Continuing Phase 1. StoredRegex.cs was already fully pure, so it moved
as-is. PatternItem.cs needed a real split: the class/PatternKind enum are
pure, but its GetAll()/GetByName() static methods are wired to
AppUtilities.TextGrabSettingsService (the settings singleton).

- Text-Grab.Core keeps the PatternItem class (properties + internal
  constructors + PatternKind enum) - no settings dependency.
- New Text-Grab/Models/PatternItemCatalog.cs (staying app-side) carries
  GetAll()/GetByName(), unchanged apart from the new name.
- Added InternalsVisibleTo("Text-Grab") and InternalsVisibleTo("Tests")
  to Text-Grab.Core so PatternItemCatalog and RegexManager.xaml.cs (which
  also constructs PatternItem directly) can still reach the internal
  constructors, and so Tests/PatternExecutorTests.cs (which constructs
  PatternItem directly for deterministic fixtures) still compiles.
- Updated every PatternItem.GetAll()/GetByName() call site to
  PatternItemCatalog: SearchBar, TextOnlyTemplateDialog,
  SplitColumnWindow, EditTextWindow, GrabFrame, and the test file.

PatternExecutor.cs stays in the app for now - this split is what
unblocks moving it next.
