+++ Text-Grab.Core/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Text-Grab")]
+[assembly: InternalsVisibleTo("Tests")]
