+++ Text-Grab.Core/Models/StoredRegex.cs
